Foodie type color and name handled
