+++ lib/presentation/bottom_nav_screens/profile_screen/review_screen/review_items.dart
@@ -32,13 +32,3 @@ class ReviewPageItems extends StatelessWidget {
     );
   }
 }
-
-Color foodietypeColor(String? foodietype) {
-  return (foodietype == '1')
-      ? AppColors.primaryGreen
-      : (foodietype == '2')
-          ? Colors.lightGreen
-          : (foodietype == '3')
-              ? Colors.lightBlue
-              : AppColors.borderGrey;
-}
